removing snapshots
