+++ packages/components/src/Chip/components/ChipPrefix/Chip.Prefix.tsx
@@ -1,9 +1,9 @@
 import React, { PropsWithChildren } from "react";
 import { useChildComponent } from "@jobber/hooks/useChildComponent";
 import classNames from "classnames";
-import styles from "./Chip.css";
-import { Avatar } from "../Avatar";
-import { Icon } from "../Icon";
+import styles from "../../Chip.css";
+import { Avatar } from "../../../Avatar";
+import { Icon } from "../../../Icon";
 
 export function ChipPrefix({ children }: PropsWithChildren) {
   const singleChild = useChildComponent(
